+++ wmipa/enumeration/__init__.py
@@ -1,4 +1,4 @@
 from .enumerator import Enumerator
 from .mathsat import MathSATEnumerator
 from .z3 import Z3Enumerator
-from .async_enumerator import AsyncEnumerator
+from .asynchronous import AsyncEnumerator
